pyproject.toml
[project]
name = "nonebot-plugin-bilibili-notifier"
version = "0.0.2"
description = "NoneBot2 plugin for Bilibili update and live notifications"
authors = [
    {name = "Variante", email = "382086919@qq.com"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0rc2", "nonebot-adapter-onebot>=2.1.5", "nonebot-plugin-apscheduler>=0.2.0", "bilibili-api-python>=16.2.0"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/Variante/nonebot-plugin-bilibili-notifier"
Repository = "https://github.com/Variante/nonebot-plugin-bilibili-notifier"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"